Port src/marks/{auto,cell,crosshair,tip} from JS to TS (stage 1)
Inline hand-written .d.ts types into the .js sources, producing single
.ts files for the auto, cell, crosshair, and tip marks. No runtime
changes; types/JSDoc come from the .d.ts and runtime from the .js.

Some imports are annotated with @ts-ignore because the shared module
.d.ts files (channel, context, options, style, text, axis, pointer,
bar) declare only a subset of the symbols actually exported by the
matching .js files; surfacing those will be addressed in a later stage
when those modules are themselves ported.
